tests: Move Eloquent relationship tests into Repository test class
